Cast to int

+++ src/kwic.cpp
@@ -66,7 +66,7 @@ DataFrame cpp_kwic(TokensPtr xptr,
         tbb::parallel_for(tbb::blocked_range<int>(0, G), [&](tbb::blocked_range<int> r) {
             for (int g = r.begin(); g < r.end(); ++g) {
                 int h = documents[g] - 1;
-                if (h < 0 || H <= h)
+                if (h < 0 || (int)H <= h)
                     throw std::range_error("Invalid documents");
                 keyword[g] = kwic(texts[h], types, delim, pos_from[g], pos_to[g]);
                 pre[g] = kwic(texts[h], types, delim, pos_from[g] - window, pos_from[g] - 1L);
@@ -78,7 +78,7 @@ DataFrame cpp_kwic(TokensPtr xptr,
 #else
     for (int g = 0; g < G; g++) {
         int h = documents[g] - 1L;
-        if (h < 0 || H <= h)
+        if (h < 0 || (int)H <= h)
             throw std::range_error("Invalid documents");
         keyword[g] = kwic(texts[h], types, delim, pos_from[g], pos_to[g]);
         pre[g] = kwic(texts[h], types, delim, pos_from[g] - window, pos_from[g] - 1L);
